+++ core/launcher/api_container_launcher/api_container_launcher.go
@@ -6,6 +6,7 @@
 package api_container_launcher
 
 import (
+	"bytes"
 	"context"
 	"fmt"
 	"github.com/docker/go-connections/nat"
@@ -16,6 +17,7 @@ import (
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"net"
+	"time"
 )
 
 const (
@@ -25,6 +27,11 @@ const (
 	shouldPublishServicePorts = true
 
 	listenProtocol = "tcp"
+
+	maxWaitForAvailabilityRetries         = 10
+	timeBetweenWaitForAvailabilityRetries = 1 * time.Second
+
+	availabilityWaitingExecCmdSuccessExitCode = 0
 
 	// The location where the enclave data directory (on the Docker host machine) will be bind-mounted
 	//  on the API container
@@ -51,10 +58,9 @@ func (launcher ApiContainerLauncher) Launch(
 	listenPort uint16,
 	gatewayIpAddr net.IP,
 	apiContainerIpAddr net.IP,
-	otherTakenIpAddrsInEnclave []net.IP,
 	isPartitioningEnabled bool,
 	enclaveDataDirpathOnHostMachine string,
-) (string, *nat.PortBinding, error){
+) (string, *nat.PortBinding, error) {
 	enclaveObjAttrsProvider := launcher.objAttrsProvider.ForEnclave(enclaveId)
 	apiContainerAttrs := enclaveObjAttrsProvider.ForApiContainer(
 		apiContainerIpAddr,
@@ -68,12 +74,11 @@ func (launcher ApiContainerLauncher) Launch(
 		gatewayIpAddr.String(): true,
 		apiContainerIpAddr.String(): true,
 	}
-	for _, takenIp := range otherTakenIpAddrsInEnclave {
-		takenIpAddrStrSet[takenIp.String()] = true
-	}
 	argsObj, err := args.NewAPIContainerArgs(
 		containerName,
 		logLevel.String(),
+		listenPort,
+		listenProtocol,
 		enclaveId,
 		networkId,
 		subnetMask,
@@ -137,6 +142,10 @@ func (launcher ApiContainerLauncher) Launch(
 		}
 	}()
 
+	if err := waitForAvailability(ctx, launcher.dockerManager, containerId, listenPort); err != nil {
+		return "", nil, stacktrace.Propagate(err, "An error occurred waiting for the API container to become available")
+	}
+
 	hostPortBinding, found := hostPortBindings[kurtosisApiPort]
 	if !found {
 		return "", nil, stacktrace.NewError("No host port binding was found for API container port '%v' - this is very strange!", kurtosisApiPort)
@@ -144,4 +153,51 @@ func (launcher ApiContainerLauncher) Launch(
 
 	shouldDeleteContainer = false
 	return containerId, hostPortBinding, nil
+}
+
+func waitForAvailability(ctx context.Context, dockerManager *docker_manager.DockerManager, containerId string, listenPortNum uint16) error {
+	commandStr := fmt.Sprintf(
+		"[ -n \"$(netstat -anp %v | grep LISTEN | grep %v)\" ]",
+		listenProtocol,
+		listenPortNum,
+	)
+	execCmd := []string{
+		"sh",
+		"-c",
+		commandStr,
+	}
+	for i := 0; i < maxWaitForAvailabilityRetries; i++ {
+		outputBuffer := &bytes.Buffer{}
+		exitCode, err := dockerManager.RunExecCommand(ctx, containerId, execCmd, outputBuffer)
+		if err == nil {
+			if (exitCode == availabilityWaitingExecCmdSuccessExitCode) {
+				return nil
+			}
+			logrus.Debugf(
+				"API container availability-waiting command '%v' returned without a Docker error, but exited with non-%v exit code '%v' and logs:\n%v",
+				commandStr,
+				availabilityWaitingExecCmdSuccessExitCode,
+				exitCode,
+				outputBuffer.String(),
+			)
+		} else {
+			logrus.Debugf(
+				"API container availability-waiting command '%v' experienced a Docker error:\n%v",
+				commandStr,
+				err,
+			)
+		}
+
+		// Tiny optimization to not sleep if we're not going to run the loop again
+		if i < maxWaitForAvailabilityRetries {
+			time.Sleep(timeBetweenWaitForAvailabilityRetries)
+		}
+	}
+
+	return stacktrace.NewError(
+		"The API container didn't become available (as measured by the command '%v') even after retrying %v times with %v between retries",
+		commandStr,
+		maxWaitForAvailabilityRetries,
+		timeBetweenWaitForAvailabilityRetries,
+	)
 }