+++ core/docs/changelog.md
@@ -1,4 +1,9 @@
 # TBD
+### Removals
+* Removed constants from the API for the listen port & listen protocol
+
+### Breaking Changes
+* Removed the API container listen port & listen protocol constants (these are now dynamically set, rather than being a global variable)
 
 # 1.28.6
 ### Features